Changed Hub so that editor is main page
