--- packages/hub/src/app/users/[username]/models/[slug]/edit/page.tsx
@@ -1,17 +1,0 @@
-"use client";
-
-import { useModelPageQuery } from "../ModelPage";
-import { EditModelPageBody } from "./EditModelPageBody";
-
-export default function Page({
-  params,
-}: {
-  params: { username: string; slug: string };
-}) {
-  const modelRef = useModelPageQuery({
-    ownerUsername: params.username,
-    slug: params.slug,
-  });
-
-  return <EditModelPageBody modelRef={modelRef} />;
-}
