+++ packages/hub/src/app/users/[username]/models/[slug]/page.tsx
@@ -1,13 +1,9 @@
 "use client";
 
-import { useFragment } from "react-relay";
+import { useModelPageQuery } from "./ModelPage";
+import { EditModelPageBody } from "./EditModelPageBody";
 
-import { ModelPage$key } from "@/__generated__/ModelPage.graphql";
-import { ModelPageFragment, useModelPageQuery } from "./ModelPage";
-import { ViewModelRevision } from "./ViewModelRevision";
-import { ViewModelRevisionContent } from "./ViewModelRevisionContent";
-
-export default function OuterModelPage({
+export default function Page({
   params,
 }: {
   params: { username: string; slug: string };
@@ -17,15 +13,5 @@ export default function OuterModelPage({
     slug: params.slug,
   });
 
-  const model = useFragment<ModelPage$key>(ModelPageFragment, modelRef);
-
-  return (
-    <ViewModelRevision
-      revisionRef={model.currentRevision}
-      modelUsername={params.username}
-      modelSlug={params.slug}
-    >
-      <ViewModelRevisionContent revisionRef={model.currentRevision} />
-    </ViewModelRevision>
-  );
+  return <EditModelPageBody modelRef={modelRef} />;
 }
